Added Item and ItemTest classes, implementing ISell

+++ src/main/java/Guitar.java
@@ -33,7 +33,7 @@ public class Guitar extends Instrument implements IPlay, ISell {
         return this.sellPrice;
     }
 
-    public void setSellPrice(int newSellPrice) {
+    public void setSellPrice(double newSellPrice) {
         this.sellPrice = newSellPrice;
     }
